libs/mng_changeling_chat/pyproject.toml
[project]
name = "mng-changeling-chat"
version = "0.1.0"
description = "Chat command plugin for mng - connect to changeling chat sessions"
requires-python = ">=3.11"
dependencies = [
    "mng==0.1.5",
    "click-option-group>=0.5.6",
]

[project.entry-points.mng]
changeling_chat = "imbue.mng_changeling_chat.plugin"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["imbue"]

[tool.uv.sources]
mng = { workspace = true }

[tool.pyright]
venvPath = "../.."
venv = ".venv"
pythonVersion = "3.11"
strict = ["**/*.py"]

[tool.pytest.ini_options]
testpaths = ["imbue/mng_changeling_chat"]
python_files = ["*_test.py", "test_*.py"]
python_classes = []
python_functions = ["test_*"]
addopts = [
  "--tb=short",
  "--strict-markers",
  "--timeout=120",
  "--cov=imbue.mng_changeling_chat",
  "--cov-report=term-missing",
  "--cov-fail-under=80",
  "-m not acceptance and not release",
]
markers = [
    "acceptance: acceptance tests (run in CI)",
    "release: release tests (run in CI)",
]

[tool.coverage.run]
source = ["imbue/mng_changeling_chat"]

[tool.coverage.report]
fail_under = 80
omit = [
    "*_test.py",
    "test_*.py",
    "conftest.py",
]
exclude_lines = [
    "pragma: no cover",
]

[[tool.importlinter.contracts]]
name = "mng_changeling_chat layers contract"
type = "layers"
layers = [
    "imbue.mng_changeling_chat.plugin",
    "imbue.mng_changeling_chat.cli",
    "imbue.mng_changeling_chat.api",
]
